update snaps and tests
